Draw highlights more safely, do not draw when not possible
Numerous changes here for safety's sake. Chiefly, we detect when the
trick is going to fail (when 'Normal' does not have a background set)
and don't try to work in that event.

Use list configuration instead of highlight configuration to know
whether we have just entered or left Visual mode. (I could probably use
some sort of flag here to be safer: to guard against the event that the
user happens to use the exact same options in Visual mode as in
not-Visual mode.)

Set the background using special keywords instead of trying to parse it
out and guess it. Only set the kind of background needed for the kind of
environment I'm in.

Try to recover from errors.